+++ docs/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased](https://github.com/LostArtefacts/TRX/compare/trx-1.0.3...develop) - ××××-××-××
 - added a fade-in and fade-out effect to patterned inventory backgrounds
+- added the ability to use monochrome image for inventory backgrounds
 - added the ability to control whether or not allies are hostile towards Lara via Lua (#3873)
 - added the ability to control via Lua which enemies are allies and which are ones that will fight with allies (#3873)
 - added support for 3D secret objects, and provided defaults for OG levels in TR2 (#4380)
@@ -57,6 +58,7 @@
 - fixed the monk in the Diving Area cutscene not having a complete death animation
 
 **TR3**:
+- added monochrome inventory backgrounds
 - added high-resolutions 16:9 and 4:3 loading screens
 - added high-resolutions 16:9 and 4:3 title and game end screens
 - added high-resolutions 16:9 and 4:3 credit images